refactored to use cobra library, updated license
