+++ cobra.yaml
@@ -1,0 +1,9 @@
+author: |
+  Vigilant Cyber Systems, Inc. 
+  Sean Heath
+  <sheath@vigilantsys.com>
+  Marc Bohler
+  <mbohler@vigilantsys.com>
+  Dylan Harbaugh
+  <dharbaugh@vigilantsys.com>
+license: MIT
