--- flake.nix
@@ -1,15 +1,0 @@
-{
-    inputs = {
-        nixpkgs.url = "github:nixos/nixpkgs";
-        flake-utils.url = "github:numtide/flake-utils";
-    };
-    outputs = { self, nixpkgs, flake-utils }:
-        flake-utils.lib.eachDefaultSystem (system:
-            let pkgs = nixpkgs.legacyPackages.${system}; in {
-                devShell = with pkgs; mkShell { buildInputs = [
-                    go_1_17
-                    gnumake
-                ];};
-            }
-        );
-}
