+++ go.mod
@@ -4,5 +4,5 @@ go 1.16
 
 require (
 	github.com/google/subcommands v1.2.0
-	github.com/spf13/cobra v1.4.0 // indirect
+	github.com/spf13/cobra v1.4.0
 )
